Add files via upload
Status Report:
Team 8 is waiting on the multiplexer for the 7-segment display to arrive. Once that gets here, we can create the countdown timer. In the meantime, we will write pseudo-code for the entire project and Arduino code for the other features we can get working without the display. So far, we know how to get input from a button and output to the LEDs, 7-segment display, and piezo buzzer.